this is a tree of all possible cases that can happen for rounding in mpfr_sub (file sub1.c) for rounding to nearest.
The nodes at depth 1 of the tree correspond to the 3 possible cases before
rounding: (1) sh=0, (2) sh>0 and the low sh bits are 1/2 ulp, (3) sh>0 and
the low sh bits are 0.

The nodes at depth 2 represent the possible subcases for k=0 in the rounding
loop.

The nodes at depth 3 represent the possible subcases for k=1 in the rounding
loop.


git-svn-id: https://scm.gforge.inria.fr/anonscm/svn/mpfr/trunk@7220 280ebfd0-de03-0410-8827-d642c229c3f4