Implemented range reduction for mpfr_sinu (and defined mpfr_fmod_ui).
This is done by computing x mod u; and a new function mpfr_fmod_ui
has been defined for this purpose. As mpfr_fmod_ui could be useful
to the user, its declaration is put in src/mpfr.h, with the intent
to document this new function later.

In the src directory:
  * Added fmod_ui.c file (mpfr_fmod_ui function) based on sqrt_ui.c
    (hence the same copyright line).
  * Makefile.am: added fmod_ui.c to libmpfr_la_SOURCES.
  * mpfr.h: added mpfr_fmod_ui.
  * sinu.c: implemented range reduction for mpfr_sinu.

git-svn-id: https://scm.gforge.inria.fr/anonscm/svn/mpfr/trunk@14242 280ebfd0-de03-0410-8827-d642c229c3f4